refactor: move shared edge modules (export, inherits, reference) to host/language_adapter
